chore: run prettier in backend

+++ backend/src/graphql/resolvers/mutations/lifetracker/deleteLtPlayer.js
@@ -1,7 +1,5 @@
 export default async (_, { name }, { user: { id: userId }, db }) => {
-  await db('ltPlayers')
-    .where({ name, userId })
-    .delete();
+  await db('ltPlayers').where({ name, userId }).delete();
 
   return true;
 };
